Commit_Diplom-2_Denisiuk_Vitalii
