+++ src/test/java/CheckAuthorizationUserTest.java
@@ -1,3 +1,4 @@
+import com.CreateUserForAuthorization;
 import com.UserForCheckAuthorization;
 import com.UserForRegistration;
 import com.UserOperations;
@@ -7,7 +8,7 @@ import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
 
-public class CheckAuthorizationUserWhoWasRegisteredTest {
+public class CheckAuthorizationUserTest {
     private UserOperations userOperations;
     private UserForCheckAuthorization userForCheckAuthorization;
 
@@ -32,5 +33,15 @@ public class CheckAuthorizationUserWhoWasRegisteredTest {
 
         Assert.assertEquals("User can not log in", 200, userAuthorizationResponse.statusCode());
         Assert.assertEquals("Wrong error message", "true", userAuthorizationResponse.jsonPath().getString("success"));
+    }
+
+    @Test
+    public void checkAuthorizationWithIncorrectEmailAndPasswordTest() {
+        CreateUserForAuthorization createUserForAuthorization = new CreateUserForAuthorization("Alex@yandex.ru", "Petrov123");
+        Response createUserForAuthorizationResponse = userOperations.userWithIncorrectDataLogin(createUserForAuthorization);
+
+        Assert.assertEquals("User can log in", 401, createUserForAuthorizationResponse.statusCode());
+        Assert.assertEquals("Wrong error message", "false", createUserForAuthorizationResponse.jsonPath().getString("success"));
+        Assert.assertEquals("Wrong error message", "email or password are incorrect", createUserForAuthorizationResponse.jsonPath().getString("message"));
     }
 }
