+++ src/test/java/CreateOrderByAuthorizedUserTest.java
@@ -6,11 +6,13 @@ import org.junit.Before;
 import org.junit.Test;
 
 public class CreateOrderByAuthorizedUserTest {
+    private OrderOperations orderOperations;
     private UserOperations userOperations;
     private IngredientsForOrder ingredientsForOrder;
 
     @Before
     public void setUp() {
+        orderOperations = new OrderOperations();
         userOperations = new UserOperations();
     }
 
@@ -29,7 +31,7 @@ public class CreateOrderByAuthorizedUserTest {
         Response userRegistrationResponse = userOperations.create(userForRegistration);
         String TokenUser = userRegistrationResponse.jsonPath().get("accessToken");
         Response userAuthorizationResponse = userOperations.userLogIn(UserForCheckAuthorization.from(userForRegistration));
-        Response checkCreateOrderResponse = userOperations.createOrderWithIngredients(ingredientsForOrder);
+        Response checkCreateOrderResponse = orderOperations.createOrderWithIngredients(ingredientsForOrder);
 
         Assert.assertEquals("Order was not created", 200, checkCreateOrderResponse.statusCode());
         Assert.assertEquals("Wrong error message", "true", checkCreateOrderResponse.jsonPath().getString("success"));
@@ -41,7 +43,7 @@ public class CreateOrderByAuthorizedUserTest {
         Response userRegistrationResponse = userOperations.create(userForRegistration);
         String TokenUser = userRegistrationResponse.jsonPath().get("accessToken");
         Response userAuthorizationResponse = userOperations.userLogIn(UserForCheckAuthorization.from(userForRegistration));
-        Response orderAuthorizedUserWithoutIngredientsResponse = userOperations.createOrder();
+        Response orderAuthorizedUserWithoutIngredientsResponse = orderOperations.createOrder();
 
         Assert.assertEquals("Order without ingredients was created", 400, orderAuthorizedUserWithoutIngredientsResponse.statusCode());
         Assert.assertEquals("Wrong error message", "false", orderAuthorizedUserWithoutIngredientsResponse.jsonPath().getString("success"));
@@ -55,7 +57,7 @@ public class CreateOrderByAuthorizedUserTest {
         Response userRegistrationResponse = userOperations.create(userForRegistration);
         String TokenUser = userRegistrationResponse.jsonPath().get("accessToken");
         Response userAuthorizationResponse = userOperations.userLogIn(UserForCheckAuthorization.from(userForRegistration));
-        Response checkCreateOrderResponse = userOperations.createOrderWithInccorectHashIngredients(ingredientsForOrderIncorrectHash);
+        Response checkCreateOrderResponse = orderOperations.createOrderWithInccorectHashIngredients(ingredientsForOrderIncorrectHash);
 
         Assert.assertEquals("Order was created", 500, checkCreateOrderResponse.statusCode());
     }
